tareas propias y colaboraciones
endpoints para tareas propias, subtareas y colaboraciones. también se restringe la edición en tareas de otros

+++ src/repositories/TareasRepository.js
@@ -3,31 +3,45 @@ class TareasRepository {
   constructor({ DBPool }) {
     this.DBPool = DBPool;
   }
+
+  async create(tarea) {
+    const { name, description, status, priority, parent_task_id, created_by } = tarea;
+    const [result] = await this.DBPool.query(
+      'INSERT INTO tasks (name, description, status, priority, parent_task_id, created_by) VALUES (?, ?, ?, ?, ?, ?)',
+      [name, description, status, priority, parent_task_id, created_by]
+    );
+    return { task_id: result.insertId, ...tarea };
+  }
+
   async findAll() {
     const [rows] = await this.DBPool.query('SELECT task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks');
     return rows;
   }
-  async findParentTasks() {
-    const [rows] = await this.DBPool.query('SELECT task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks WHERE parent_task_id is null');
+  async getParentTasksByUserID(user_id) {
+    const [rows] = await this.DBPool.query('SELECT task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks WHERE parent_task_id is null and created_by = ?', [user_id]);
     return rows;
   }
   async findSubTasks(parent_task_id) {
     const [rows] = await this.DBPool.query('SELECT task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks WHERE parent_task_id = ?', [parent_task_id]);
     return rows;
   }
+  async findTasksByColaborating(user_id) {
+    const [rows] = await this.DBPool.query('SELECT tasks.task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks join user_tasks ON tasks.task_id = user_tasks.task_id WHERE user_tasks.user_id = ? and parent_task_id is null', [user_id]);
+    return rows;
+  }
   async findById(id) {
-    const [rows] = await this.DBPool.query('SELECT task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks WHERE task_id = ?', [id]);
+    const [rows] = await this.DBPool.query('SELECT created_by, task_id, name, description, status, priority, created_at, updated_at, parent_task_id FROM tasks WHERE task_id = ?', [id]);
     return rows.length ? rows[0] : null;
   }
   async findByOwnerId(owner_id) {
     const [rows] = await this.DBPool.query('SELECT * FROM tasks WHERE created_by = ?', [owner_id]);
     return rows.length ? rows[0] : null;
   }
-  
   async findByParentId(parent_id) {
     const [rows] = await this.DBPool.query('SELECT * FROM tasks WHERE parent_task_id = ?', [parent_id]);
     return rows.length ? rows[0] : null;
   }
+
   async findPriorities() {
     const [rows] = await this.DBPool.query("SHOW COLUMNS FROM tasks LIKE 'priority'");
     
@@ -38,7 +52,6 @@ class TareasRepository {
     
     return enumValues;
   }
-  
   async findStatus() {
     const [rows] = await this.DBPool.query("SHOW COLUMNS FROM tasks LIKE 'status'");
     
@@ -50,26 +63,17 @@ class TareasRepository {
     return enumValues;
   }
 
-  async create(tarea) {
-    const { name, description, status, priority, parent_task_id, created_by } = tarea;
-    const [result] = await this.DBPool.query(
-      'INSERT INTO tasks (name, description, status, priority, parent_task_id, created_by) VALUES (?, ?, ?, ?, ?, ?)',
-      [name, description, status, priority, parent_task_id, created_by]
-    );
-    return { task_id: result.insertId, ...tarea };
-  }
-
-  async update(id, tarea) {
+  async update(id, tarea, user_id) {
     const { name, description, status, priority, parent_task_id} = tarea;
     const [result] = await this.DBPool.query(
-      'UPDATE tasks SET name = ?, description = ?, status = ?, priority = ?, parent_task_id = ?, updated_at = NOW() WHERE task_id = ?',
-      [name, description, status, priority, parent_task_id, id]
+      'UPDATE tasks SET name = ?, description = ?, status = ?, priority = ?, parent_task_id = ?, updated_at = NOW() WHERE task_id = ? and created_by = ?',
+      [name, description, status, priority, parent_task_id, id, user_id]
     );
     return result.affectedRows > 0 ? { task_id: id, ...tarea } : null;
   }
 
-  async delete(id) {
-    const [result] = await this.DBPool.query('DELETE FROM tasks WHERE task_id = ?', [id]);
+  async delete(id, user_id) {
+    const [result] = await this.DBPool.query('DELETE FROM tasks WHERE task_id = ? and created_by = ?', [id, user_id]);
     return result.affectedRows > 0;
   }
 }
